Convert project from IntelliJ plugin into Gradle (#28)
* Remove compiled artifacts from the repo

It will be possible to re-create thsese artifacts via a single Gradle
command.

* Convert project from IntelliJ plugin into Gradle

The new project model is based on Gradle and [gradle-intellij-plugin][1]
which is now the preferred way of building and running plugins for the
IntelliJ platform.

The project model in Gradle is defined via the build.gradle file and
available via the command line via the `./gradlew` executable from the
project directory.

You can build the plugin via `./gradlew buildPlugin`, run the IDE with this
plugin set up in a sandbox via `./gradlew runIde`, and so on. See the repo
for gradle-intellij-plugin.

To start developing the plugin in IntelliJ you should open the directory
with the plugin via the "File | Open" menu. Gradle configuration is
located in the "Gradle" tool window via "View | Tool Windows | Gradle".
You can find all the Gradle tasks there, including "clean", "runIde",
"buildPlugin", "check", etc. Thre is also the action "Refresh all
Gradle projects" (the icon with two refresh/sync arrows) that allows to
sync your project with any changes in your build.gradle file.

  [1]: https://github.com/JetBrains/gradle-intellij-plugin